requirements.txt
azure-storage-blob==12.11.0
azure-core==1.23.1
azure-common==1.1.28
azure-storage-common==2.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
greenlet==1.1.2
idna==3.3
isodate==0.6.1
msrest==0.6.21
numpy==1.22.3
oauthlib==3.2.0
pandas==1.4.2
pip==22.0.4
py4j==0.10.9.5
pyarrow==7.0.0
pycparser==2.21
pyodbc==4.0.32
pyspark==3.2.1
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
setuptools==62.0.0
six==1.16.0
SQLAlchemy==1.4.34
typing-extensions==4.1.1
urllib3==1.26.9
wheel==0.37.1
yml==0.0.1